Revert "support multiple private keys"
turns out this is useless/impossible

This reverts commit 0abc0610ec9c4c25466c8cadb8edbbed01ce17e9.

+++ assets/common.sh
@@ -1,20 +1,17 @@
 export TMPDIR=${TMPDIR:-/tmp}
 
 load_pubkey() {
-  local private_keys_path=$TMPDIR/git-resource-private-keys
+  local private_key_path=$TMPDIR/git-resource-private-key
 
-  (jq '([.source.private_key // empty] + (.source.private_keys // []))[]' < $1) > $private_keys_path
+  (jq -r '.source.private_key // empty' < $1) > $private_key_path
 
-  eval $(ssh-agent) >/dev/null 2>&1
-  trap "kill $SSH_AGENT_PID" 0
+  if [ -s $private_key_path ]; then
+    chmod 0600 $private_key_path
 
-  while IFS= read -r key; do
-    keyfile=$(mktemp $TMPDIR/key.XXXXXX)
+    eval $(ssh-agent) >/dev/null 2>&1
+    trap "kill $SSH_AGENT_PID" 0
 
-    echo "$key" | jq -r . > $keyfile
-    chmod 0600 $keyfile
-
-    SSH_ASKPASS=/opt/resource/askpass.sh DISPLAY= ssh-add $keyfile >/dev/null
+    SSH_ASKPASS=/opt/resource/askpass.sh DISPLAY= ssh-add $private_key_path >/dev/null
 
     mkdir -p ~/.ssh
     cat > ~/.ssh/config <<EOF
@@ -22,7 +19,7 @@ StrictHostKeyChecking no
 LogLevel quiet
 EOF
     chmod 0600 ~/.ssh/config
-  done < $private_keys_path
+  fi
 }
 
 git_metadata() {
